Reorganize core module
Move config and multi-lang to separate folders.
Create a new `ml` ES sub-module for multi-lang.
Rename functions so the intended usage becomes
- qTranx.ml.splitLangs
- qTranx.ml.splitTokens
- qTranx.ml.parseTokens

Fix external usage in dev yoast scripts.
