+++ dev/wp-seo-yoast-13/js/qwpseo-prep.js
@@ -19,7 +19,7 @@
             return;
 
         //Yoast will delete this field in term-scraper
-        var contents = mlSplitLangs(d.val());
+        var contents = qtranx.ml.splitLangs(d.val());
         h.contentField.value = contents[qTranx.hooks.getActiveLanguage()];
         for (var lang in h.fields) {
             h.fields[lang].value = contents[lang];
