--- js/core/index.js
@@ -1,7 +1,0 @@
-/**
- * Shared core functionalities and configuration, part of the public API.
- */
-'use strict';
-
-export * from './config';
-export * from './multi-lang';
