+++ js/hooks/deprecated.js
@@ -3,7 +3,7 @@
  * ATTENTION - ALL THESE FUNCTIONS WILL BE REMOVED IN NEXT MAJOR RELEASE.
  */
 'use strict';
-import {config} from '../core/config';
+import {config} from '../config';
 import {addContentHook, addContentHooks, addDisplayHook, addDisplayHooks} from './handlers';
 
 // -------------------------------------------------------------------------------------------------------
@@ -50,7 +50,7 @@ export const getFlagLocation = function () {
  *
  * @param {string} lang
  * @returns {boolean} true if 'lang' is in the hash of enabled languages.
- * This function maybe needed, as function mlSplitLangs may return languages,
+ * This function maybe needed, as function ml.splitLangs may return languages,
  * which are not enabled, in case they were previously enabled and had some data.
  * Such data is preserved and re-saved until user deletes it manually.
  */
