+++ js/multi-lang/index.js
@@ -1,0 +1,6 @@
+/**
+ * Package for multi-lang text.
+ */
+'use strict';
+
+export * from './parser';
